move multiterm query to a separate translation unit
