+++ core/search/granular_range_filter.cpp
@@ -25,7 +25,7 @@
 
 #include "boolean_filter.hpp"
 #include "range_filter.hpp"
-#include "range_query.hpp"
+#include "multiterm_query.hpp"
 #include "term_query.hpp"
 #include "analysis/token_attributes.hpp"
 #include "index/index_reader.hpp"
@@ -50,7 +50,7 @@ NS_LOCAL
 // max_term (with e.g. N=3)-/                                                 
 //////////////////////////////////////////////////////////////////////////////
 
-typedef std::unordered_multimap<const irs::sub_reader*, irs::range_state> granular_states_t;
+typedef std::unordered_multimap<const irs::sub_reader*, irs::multiterm_state> granular_states_t;
 
 // return the granularity portion of the term
 irs::bytes_ref mask_granularity(const irs::bytes_ref& term, size_t prefix_size) {
@@ -72,7 +72,7 @@ irs::bytes_ref mask_value(const irs::bytes_ref& term, size_t prefix_size) {
 
 // collect terms while they are accepted by Comparer
 template<typename Comparer>
-irs::range_state& collect_terms(
+irs::multiterm_state& collect_terms(
     granular_states_t& states,
     const irs::sub_reader& reader,
     const irs::term_reader& tr,
@@ -88,8 +88,6 @@ irs::range_state& collect_terms(
   // initialize range state
   terms.read(); // read attributes (needed for cookie())
   state.reader = &tr;
-  state.min_term = terms.value();
-  state.min_cookie = terms.cookie();
 
   auto& meta = terms.attributes().get<irs::term_meta>(); // get term metadata
   const decltype(irs::term_meta::docs_count) NO_DOCS = 0;
@@ -102,16 +100,14 @@ irs::range_state& collect_terms(
       break; // terminate traversal
     }
 
-
     // fill scoring candidates
     scorer.collect(
       docs_count,
-      state.count, // current term offset in state
+      state.count++, // current term offset in state
       state,
       reader,
       terms
     );
-    ++(state.count);
 
     state.estimation += docs_count;
   } while (terms.next());
@@ -593,7 +589,7 @@ filter::prepared::ptr by_granular_range::prepare(
   // group the range states into a minimal number of groups per sub_reader
   // ...........................................................................
 
-  std::vector<range_query::states_t> range_states;
+  std::vector<multiterm_query::states_t> range_states;
   size_t current_states = 0;
   const sub_reader* previous_reader = nullptr;
 
@@ -623,24 +619,24 @@ filter::prepared::ptr by_granular_range::prepare(
   // ...........................................................................
 
   // dummy class for returning the stored prepared query on a call to prepare(...)
-  class range_filter_proxy: public filter {
+  class multiterm_filter_proxy: public filter {
    public:
-    range_query::ptr query_;
-    range_filter_proxy(): filter(by_range::type()) {}
-    static ptr make() { return memory::make_unique<range_filter_proxy>(); }
+    multiterm_query::ptr query_;
+    multiterm_filter_proxy(): filter(by_range::type()) {}
+    static ptr make() { return memory::make_unique<multiterm_filter_proxy>(); }
     virtual filter::prepared::ptr prepare(
       const index_reader&,
       const order::prepared&,
       boost_t,
       const attribute_view&) const override {
-      return query_;
+        return query_;
     }
   };
 
   Or multirange_filter;
 
   for (auto& range_state: range_states) {
-    multirange_filter.add<range_filter_proxy>().query_ = memory::make_shared<range_query>(std::move(range_state), std::move(stats), irs::no_boost());
+    multirange_filter.add<multiterm_filter_proxy>().query_ = memory::make_shared<multiterm_query>(std::move(range_state), std::move(stats), irs::no_boost());
   }
 
   return multirange_filter.boost(this->boost()).prepare(rdr, ord, boost);
@@ -660,20 +656,22 @@ bool by_granular_range::equals(const filter& rhs) const noexcept {
 // -----------------------------------------------------------------------------
 
 bstring& by_granular_range::insert(
-  terms_t& terms, const level_t& granularity_level
-) {
+    terms_t& terms,
+    const level_t& granularity_level) {
   return terms[granularity_level];
 }
 
 bstring& by_granular_range::insert(
-  terms_t& terms, const level_t& granularity_level, bstring&& term
-) {
+    terms_t& terms,
+    const level_t& granularity_level,
+    bstring&& term) {
   return terms[granularity_level] = std::move(term);
 }
 
 bstring& by_granular_range::insert(
-  terms_t& terms, const level_t& granularity_level, const bytes_ref& term
-) {
+    terms_t& terms,
+    const level_t& granularity_level,
+    const bytes_ref& term) {
   return terms[granularity_level] = term;
 }
