+++ core/search/multiterm_query.hpp
@@ -1,0 +1,70 @@
+////////////////////////////////////////////////////////////////////////////////
+/// DISCLAIMER
+///
+/// Copyright 2019 ArangoDB GmbH, Cologne, Germany
+///
+/// Licensed under the Apache License, Version 2.0 (the "License");
+/// you may not use this file except in compliance with the License.
+/// You may obtain a copy of the License at
+///
+///     http://www.apache.org/licenses/LICENSE-2.0
+///
+/// Unless required by applicable law or agreed to in writing, software
+/// distributed under the License is distributed on an "AS IS" BASIS,
+/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and
+/// limitations under the License.
+///
+/// Copyright holder is ArangoDB GmbH, Cologne, Germany
+///
+/// @author Andrey Abramov
+////////////////////////////////////////////////////////////////////////////////
+
+#ifndef IRESEARCH_MULTITERM_QUERY_H
+#define IRESEARCH_MULTITERM_QUERY_H
+
+#include "filter.hpp"
+#include "cost.hpp"
+#include "limited_sample_scorer.hpp"
+
+NS_ROOT
+
+//////////////////////////////////////////////////////////////////////////////
+/// @struct multiterm_state
+/// @brief cached per reader range state
+//////////////////////////////////////////////////////////////////////////////
+struct multiterm_state : limited_sample_state {
+  size_t count{}; // number of matched terms in a state
+  cost::cost_t estimation{}; // estimated cost
+};
+
+//////////////////////////////////////////////////////////////////////////////
+/// @class multiterm_query
+/// @brief compiled query suitable for filters with non adjacent set of terms
+//////////////////////////////////////////////////////////////////////////////
+class multiterm_query : public filter::prepared {
+ public:
+  typedef states_cache<multiterm_state> states_t;
+  typedef std::vector<bstring> stats_t;
+
+  DECLARE_SHARED_PTR(multiterm_query);
+
+  explicit multiterm_query(states_t&& states, stats_t&& stats, boost_t boost)
+    : prepared(boost),
+      states_(std::move(states)),
+      stats_(std::move(stats)) {
+  }
+
+  virtual doc_iterator::ptr execute(
+      const sub_reader& rdr,
+      const order::prepared& ord,
+      const attribute_view& /*ctx*/) const override;
+
+ private:
+  states_t states_;
+  stats_t stats_;
+}; // multiterm_query
+
+NS_END // ROOT
+
+#endif // IRESEARCH_MULTITERM_QUERY_H
