+++ core/search/prefix_filter.cpp
@@ -21,28 +21,80 @@
 /// @author Vasiliy Nabatchikov
 ////////////////////////////////////////////////////////////////////////////////
 
-#include "shared.hpp"
 #include "prefix_filter.hpp"
-#include "range_query.hpp"
+
+#include <boost/functional/hash.hpp>
+
+#include "shared.hpp"
+#include "multiterm_query.hpp"
 #include "analysis/token_attributes.hpp"
 #include "index/index_reader.hpp"
 #include "index/iterators.hpp"
 
-#include <boost/functional/hash.hpp>
-
 NS_ROOT
-
-filter::prepared::ptr by_prefix::prepare(
-    const index_reader& index,
-    const order::prepared& ord,
-    boost_t boost,
-    const attribute_view& /*ctx*/) const {
-  return range_query::make_from_prefix(index, ord, this->boost()*boost,
-                                       field(), term(), scored_terms_limit_);
-}
 
 DEFINE_FILTER_TYPE(by_prefix)
 DEFINE_FACTORY_DEFAULT(by_prefix)
+
+/*static*/ filter::prepared::ptr by_prefix::prepare(
+    const index_reader& index,
+    const order::prepared& ord,
+    boost_t boost,
+    const string_ref& field,
+    const bytes_ref& prefix,
+    size_t scored_terms_limit) {
+  limited_sample_scorer scorer(ord.empty() ? 0 : scored_terms_limit); // object for collecting order stats
+  multiterm_query::states_t states(index.size());
+
+  // iterate over the segments
+  for (const auto& segment: index) {
+    // get term dictionary for field
+    const term_reader* reader = segment.field(field);
+
+    if (!reader) {
+      continue;
+    }
+
+    seek_term_iterator::ptr terms = reader->iterator();
+
+    // seek to prefix
+    if (SeekResult::END == terms->seek_ge(prefix)) {
+      continue;
+    }
+
+    auto& value = terms->value();
+
+    // get term metadata
+    auto& meta = terms->attributes().get<term_meta>();
+    const decltype(irs::term_meta::docs_count) NO_DOCS = 0;
+    const auto& docs_count = meta ? meta->docs_count : NO_DOCS;
+
+    if (starts_with(value, prefix)) {
+      terms->read();
+
+      // get state for current segment
+      auto& state = states.insert(segment);
+      state.reader = reader;
+
+      do {
+        // fill scoring candidates
+        scorer.collect(docs_count, state.count++, state, segment, *terms);
+        state.estimation += docs_count; // collect cost
+
+        if (!terms->next()) {
+          break;
+        }
+
+        terms->read();
+      } while (starts_with(value, prefix));
+    }
+  }
+
+  std::vector<bstring> stats;
+  scorer.score(index, ord, stats);
+
+  return memory::make_shared<multiterm_query>(std::move(states), std::move(stats), boost);
+}
 
 by_prefix::by_prefix() noexcept : by_prefix(by_prefix::type()) { }
