+++ core/search/range_filter.cpp
@@ -21,14 +21,57 @@
 /// @author Vasiliy Nabatchikov
 ////////////////////////////////////////////////////////////////////////////////
 
-#include "shared.hpp"
 #include "range_filter.hpp"
-#include "range_query.hpp"
+
+#include <boost/functional/hash.hpp>
+
+#include "shared.hpp"
+#include "multiterm_query.hpp"
 #include "term_query.hpp"
 #include "index/index_reader.hpp"
 #include "analysis/token_attributes.hpp"
 
-#include <boost/functional/hash.hpp>
+NS_LOCAL
+
+template<typename Comparer>
+void collect_terms(
+    const irs::sub_reader& segment,
+    const irs::term_reader& field,
+    irs::seek_term_iterator& terms,
+    irs::multiterm_query::states_t& states,
+    irs::limited_sample_scorer& scorer,
+    Comparer cmp) {
+  auto& value = terms.value();
+
+  if (cmp(value)) {
+    // read attributes
+    terms.read();
+
+    // get state for current segment
+    auto& state = states.insert(segment);
+    state.reader = &field;
+
+    // get term metadata
+    auto& meta = terms.attributes().get<irs::term_meta>();
+    const decltype(irs::term_meta::docs_count) NO_DOCS = 0;
+    const auto& docs_count = meta ? meta->docs_count : NO_DOCS;
+
+    do {
+      // fill scoring candidates
+      scorer.collect(docs_count, state.count++, state, segment, terms);
+      state.estimation += docs_count;
+
+      if (!terms.next()) {
+        break;
+      }
+
+      // read attributes
+      terms.read();
+    } while (cmp(value));
+  }
+}
+
+NS_END
 
 NS_ROOT
 
@@ -60,6 +103,7 @@ filter::prepared::ptr by_range::prepare(
     const order::prepared& ord,
     boost_t boost,
     const attribute_view& /*ctx*/) const {
+  boost *= this->boost();
   //TODO: optimize unordered case
   // - seek to min
   // - get ordinal position of the term
@@ -72,14 +116,78 @@ filter::prepared::ptr by_range::prepare(
 
     if (rng_.min_type == rng_.max_type && rng_.min_type == BoundType::INCLUSIVE) {
       // degenerated case
-      return term_query::make(index, ord, boost*this->boost(), fld_, rng_.min);
+      return term_query::make(index, ord, boost, fld_, rng_.min);
     }
 
     // can't satisfy conditon
     return prepared::empty();
   }
 
-  return range_query::make_from_range(index, ord, this->boost()*boost, fld_, rng_, scored_terms_limit_);
+  limited_sample_scorer scorer(ord.empty() ? 0 : scored_terms_limit()); // object for collecting order stats
+  multiterm_query::states_t states(index.size());
+
+  const string_ref field = this->field();
+
+  // iterate over the segments
+  for (const auto& segment : index) {
+    // get term dictionary for field
+    const auto* reader = segment.field(field);
+
+    if (!reader) {
+      // can't find field with the specified name
+      continue;
+    }
+
+    auto terms = reader->iterator();
+    bool res = false;
+
+    // seek to min
+    switch (rng_.min_type) {
+      case BoundType::UNBOUNDED:
+        res = terms->next();
+        break;
+      case BoundType::INCLUSIVE:
+        res = seek_min<true>(*terms, rng_.min);
+        break;
+      case BoundType::EXCLUSIVE:
+        res = seek_min<false>(*terms, rng_.min);
+        break;
+    }
+
+    if (!res) {
+      // reached the end, nothing to collect
+      continue;
+    }
+
+    // now we are on the target or the next term
+    const irs::bytes_ref max = rng_.max;
+
+    switch (rng_.max_type) {
+      case BoundType::UNBOUNDED:
+        ::collect_terms(
+          segment, *reader, *terms, states, scorer, [](const bytes_ref&) {
+            return true;
+        });
+        break;
+      case BoundType::INCLUSIVE:
+        ::collect_terms(
+          segment, *reader, *terms, states, scorer, [max](const bytes_ref& term) {
+            return term <= max;
+        });
+        break;
+      case BoundType::EXCLUSIVE:
+        ::collect_terms(
+          segment, *reader, *terms, states, scorer, [max](const bytes_ref& term) {
+            return term < max;
+        });
+        break;
+    }
+  }
+
+  std::vector<bstring> stats;
+  scorer.score(index, ord, stats);
+
+  return memory::make_shared<multiterm_query>(std::move(states), std::move(stats), boost);
 }
 
 NS_END // ROOT
